[update]electron用js修正
