+++ public/electron.js
@@ -1,0 +1,39 @@
+const electron = require('electron')
+const app = electron.app
+const BrowserWindow = electron.BrowserWindow
+
+const isDev = require('electron-is-dev')
+let mainWindow
+
+function createWindow () {
+  mainWindow = new BrowserWindow({
+    width: 800,
+    height: 600,
+    webPreferences: {
+      nodeIntegration: true
+    }
+  })
+  // Load file
+  mainWindow.loadFile("build/index.html")
+
+  // Open the DevTools.
+  if (isDev) {
+    mainWindow.webContents.openDevTools()
+  }
+
+  mainWindow.on('closed', () => (mainWindow = null))
+}
+
+app.on('ready', createWindow)
+
+app.on('window-all-closed', () => {
+  if (process.platform !== 'darwin') {
+    app.quit()
+  }
+})
+
+app.on('activate', () => {
+  if (mainWindow === null) {
+    createWindow()
+  }
+})
